+++ spec/rubocop/cop/style/empty_lines_around_class_body_spec.rb
@@ -191,4 +191,59 @@ describe RuboCop::Cop::Style::EmptyLinesAroundClassBody, :config do
                                 'end'].join("\n"))
     end
   end
+
+  context 'when EnforcedStyle is empty_lines' do
+    let(:cop_config) { { 'EnforcedStyle' => 'body_start_only' } }
+
+    it 'registers an offense for class body not starting with a blank' do
+      inspect_source(cop,
+                     ['class SomeClass',
+                      '  do_something',
+                      'end'])
+      expect(cop.messages).to eq(
+        ['Empty line missing at class body beginning.'])
+    end
+
+    it 'autocorrects class body containing nothing' do
+      corrected = autocorrect_source(cop,
+                                     ['class SomeClass',
+                                      'end'])
+      expect(corrected).to eq ['class SomeClass',
+                               '',
+                               'end'].join("\n")
+    end
+
+    it 'autocorrects beginning and end' do
+      new_source = autocorrect_source(cop,
+                                      ['class SomeClass',
+                                       '  do_something',
+                                       '',
+                                       'end'])
+      expect(new_source).to eq(['class SomeClass',
+                                '',
+                                '  do_something',
+                                'end'].join("\n"))
+    end
+
+    it 'registers offense for singleton class body not starting with a blank' do
+      inspect_source(cop,
+                     ['class << self',
+                      '  do_something',
+                      'end'])
+      expect(cop.messages).to eq(
+        ['Empty line missing at class body beginning.'])
+    end
+
+    it 'autocorrects beginning and end' do
+      new_source = autocorrect_source(cop,
+                                      ['class << self',
+                                       '  do_something',
+                                       '',
+                                       'end'])
+      expect(new_source).to eq(['class << self',
+                                '',
+                                '  do_something',
+                                'end'].join("\n"))
+    end
+  end
 end